backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
python-multipart
pandas
numpy
libsql-client
infisicalsdk>=1.0.0
python-dotenv
pytz
requests
toml
yfinance
